Add rights Fix Condition

+++ ajax/deletedraft.php
@@ -33,11 +33,16 @@ use GlpiPlugin\Metademands\Draft_Value;
 header("Content-Type: text/html; charset=UTF-8");
 Html::header_nocache();
 
-$users_id                          = $_POST['users_id'];
-$plugin_metademands_metademands_id = $_POST['plugin_metademands_metademands_id'];
-$draft_id                          = $_POST['drafts_id'];
+Session::checkLoginUser();
+
+$users_id                          = Session::getLoginUserID();
+$plugin_metademands_metademands_id = (int) $_POST['plugin_metademands_metademands_id'];
+$draft_id                          = (int) $_POST['drafts_id'];
 
 $self = new Draft();
+if (!$self->getFromDB($draft_id) || (int) $self->fields['users_id'] !== $users_id) {
+    throw new \Glpi\Exception\Http\AccessDeniedHttpException();
+}
 $self->deleteByCriteria(['id' => $draft_id]);
 
 $values = new Draft_Value();
